+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `cptstodt` and `cpdttots` functions to `cp.h` to work with `struct timespec`.
+
+
 ## [0.3.0] - 2025-01-16
 ### Added
 - Support for extracting unsigned SHV integer to the signed one and vice versa.